Update README.md
Bitcoin Key Utilities is a Python script designed to help users perform Bitcoin arithmetic and cryptographic operations. It is intended to be a helpful tool for users who need to work with Bitcoin public and private keys, and for those who may have lost their partial private keys but still possess the corresponding public key. Please note that this script is not meant for any illegal hacking purposes but rather for legitimate Bitcoin-related activities.